Decode lightning data
